feat: a much cleaner example
