+++ src/app/views/examples/dialog-service-example/count-down-dialog/count-down-dialog.component.html
@@ -1,0 +1,9 @@
+<forge-toolbar>
+  <h1 id="dialog-title" slot="start" class="forge-typography--title">Are you there?</h1>
+</forge-toolbar>
+<section>
+  <div class="forge-typography--body1">Looks like you've stepped away. We'll log you out in...</div>
+  <div class="seconds-container forge-typography--headline5">
+    <span id="seconds">{{counter$ | async}}</span><span>seconds</span>
+  </div>
+</section>
