+++ src/app/views/examples/dialog-service-example/count-down-dialog/count-down-dialog.component.scss
@@ -1,0 +1,18 @@
+:host {
+  display: block;
+}
+
+.seconds-container {
+  margin-top: 8px;
+  display: flex;
+  justify-content: flex-end;
+  align-items: flex-end;
+}
+
+section {
+  padding: 16px;
+
+  & span:last-child {
+    margin-left: 8px;
+  }
+}
