+++ src/app/views/examples/dialog-service-example/dialog-service-example.module.ts
@@ -3,12 +3,13 @@ import { CommonModule } from '@angular/common';
 import { DialogServiceExampleComponent } from './dialog-service-example.component';
 import { DemoCardComponent } from 'src/app/shared/components/demo-card/demo-card.component';
 import { DialogServiceExampleRoutingModule } from './dialog-service-example-routing.module';
-import { ForgeButtonModule, ForgeCheckboxModule, ForgeDividerModule, ForgeTextFieldModule} from '@tylertech/forge-angular';
+import { ForgeButtonModule, ForgeCheckboxModule, ForgeDividerModule, ForgeTextFieldModule, ForgeToolbarModule} from '@tylertech/forge-angular';
 import { DialogComponent } from './dialog/dialog.component';
 import { FormsModule } from '@angular/forms';
+import { CountDownDialogComponent } from './count-down-dialog/count-down-dialog.component';
 
 @NgModule({
-  declarations: [DialogServiceExampleComponent, DialogComponent],
+  declarations: [DialogServiceExampleComponent, DialogComponent, CountDownDialogComponent],
   imports: [
     CommonModule,
     FormsModule,
@@ -17,7 +18,8 @@ import { FormsModule } from '@angular/forms';
     ForgeButtonModule,
     ForgeTextFieldModule,
     ForgeDividerModule,
-    ForgeCheckboxModule
+    ForgeCheckboxModule,
+    ForgeToolbarModule
   ]
 })
 export class DialogServiceExampleModule {}
