+++ src/app/views/examples/dialog-service-example/dialog/dialog.component.html
@@ -1,5 +1,7 @@
-<div class="app-dialog" >
-  <h2 forge-dialog-move-target>Open dialog</h2>
-  <p>Dialog will be closed by the service in {{counter$ | async}}</p>
-</div>
+<forge-toolbar>
+  <h1 id="dialog-title" slot="start" class="forge-typography--title">Example dialog</h1>
+</forge-toolbar>
+<section>
+  <div class="forge-typography--body1">This is an example dialog that will be closed by the dialog service.</div>
+</section>
