+++ src/app/views/examples/dialog-service-example/dialog/dialog.component.scss
@@ -1,5 +1,7 @@
-.app-dialog {
+:host {
   padding: 16px;
+  width: 80dvw;
+  height: 50dvh;
 
   h2 {
     margin-top: 0;
